Refactor webview file handling and state management
- Introduced new module for file handling to improve code organization and maintainability.
- Refactored script.js to utilize new fileHandler and stateManager modules, enhancing clarity and reducing redundancy.
- Updated multiple file selection logic, including improved visibility toggling and state restoration.
- Cleaned up utility functions and constants for better readability and consistency.
- Enhanced event handling for multiple selections and output files, ensuring a smoother user experience.
