+++ coauthor-for-vs-code/src/webview/index.html
@@ -331,7 +331,10 @@
         <div class="file-select">
           <div class="file-select-header">
             <label for="multipleOutputFilesSelect">
-              <span id="toggleMultipleOutputFiles" class="toggle-icon" title="More"
+              <span
+                id="toggleMultipleOutputFiles"
+                class="toggle-icon"
+                title="More"
                 >▼</span
               ><span class="advanced-feature">Multiple File Output</span>
             </label>
